requirements.txt
django
tablib
django-import-export
django-crispy-forms